Build 60 ("the secret to comedy")
This build seems to (at least on my iPhone 15 Pro) fix some of the timing problems that were getting the metadata out of sync with which station was playing.
The activity trace that was just added seems to be showing much more frequent updates to the metadata than I expect; this may be a bug, or it may not. Needs
more debugging before we go public with 3.0.

Please test by listening and switching stations, both quickly and slowly, and try station switches from the lockscreen too.